Rename SplitSelector -> Selector
