+++ src/pages/iou/steps/MoneyRequstParticipantsPage/MoneyRequestParticipantsPage.js
@@ -6,7 +6,7 @@ import lodashGet from 'lodash/get';
 import CONST from '../../../../CONST';
 import ONYXKEYS from '../../../../ONYXKEYS';
 import ROUTES from '../../../../ROUTES';
-import MoneyRequestParticipantsSplitSelector from './MoneyRequestParticipantsSplitSelector';
+import MoneyRequestParticipantsSelector from './MoneyRequestParticipantsSelector';
 import styles from '../../../../styles/styles';
 import ScreenWrapper from '../../../../components/ScreenWrapper';
 import withLocalize, {withLocalizePropTypes} from '../../../../components/withLocalize';
@@ -67,7 +67,7 @@ function MoneyRequestParticipantsPage({iou, translate, route}) {
                         title={translate('iou.cash')}
                         onBackButtonPress={navigateBack}
                     />
-                        <MoneyRequestParticipantsSplitSelector
+                        <MoneyRequestParticipantsSelector
                             participants={iou.participants}
                             onAddParticipants={IOU.setMoneyRequestParticipants}
                             navigateToRequest={() => navigateToNextStep(CONST.IOU.MONEY_REQUEST_TYPE.REQUEST)}
